updated routing, changed stying, added folders for styles and components
