+++ src/App.js
@@ -1,37 +1,15 @@
 import "./App.css";
-import Location from "./Location";
-import WelcomePage from "./WelcomePage";
-import Header from "./Header";
-import { Link, Routes, Route } from "react-router-dom";
-import axios from "axios";
-import Results from "./Results";
-import SearchItems from "./SearchItems";
-import Footer from "./Footer";
+
+import Header from "./components/Header";
+import Main from "./components/Main";
+import Footer from "./components/Footer";
 
 function App() {
-  const apiKey = "SbABP9Vr89Ox8a38s29QPLUQm51xa784";
-
   return (
     <>
       <div className="App">
         <Header />
-        <main id="mainContent">
-          <section className="welcomePageDiv">
-            <Routes>
-              <Route path="/" element={<WelcomePage />} />
-              <Route path="/location" element={<Location apiKey={apiKey} />} />
-
-              <Route
-                path="/location/:coords"
-                element={<SearchItems apiKey={apiKey} />}
-              />
-              <Route
-                path="/location/:coords/:searchItem"
-                element={<Results apiKey={apiKey} />}
-              />
-            </Routes>
-          </section>
-        </main>
+        <Main />
         <Footer />
       </div>
     </>
