--- src/BaseStyles.css
@@ -1,26 +1,0 @@
-* {
-  --green: #c4e39c;
-  --blue: #5bc5ff;
-  --orange: #ff9d7f;
-  --black: rgb(22, 22, 22);
-  --input-grey: #e4e4e4;
-  --spacing: calc(8px + 1.5625vw);
-}
-
-body {
-  background-color: var(--orange);
-}
-
-html {
-  font-size: 125%;
-}
-
-.App {
-  height: 100vh;
-}
-
-.wrapper {
-  max-width: 1280px;
-  width: 90%;
-  margin: 0 auto;
-}
