--- src/Main.css
@@ -1,246 +1,0 @@
-/* Main section */
-
-main {
-  background-color: white;
-  border-top-left-radius: 15px;
-  border-top-right-radius: 15px;
-  height: 50vh;
-}
-
-.getStartedButton {
-  width: 50%;
-  margin: 40px auto;
-}
-
-/* LOCATION POPUP */
-.locationPopup {
-  display: none;
-  top: 0;
-  left: 0;
-  height: 100vh;
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.5);
-}
-
-.locationPopup.active {
-  display: block;
-}
-
-.locationPopupContent {
-  width: 90%;
-  margin: 20% auto;
-  padding: 20px;
-  background-color: white;
-  text-align: center;
-  border-radius: 15px;
-}
-
-.locationPopup .findLocation {
-  width: 70%;
-  margin: 0 auto;
-}
-
-.popupButtons {
-  display: flex;
-  flex-direction: column;
-}
-.closeLocation {
-  border: none;
-  background-color: transparent;
-  font-family: var(--bodyFont);
-  margin: 20px;
-}
-
-.closeLocation:hover,
-.closeLocation:focus {
-  color: brown;
-  cursor: pointer;
-}
-
-/* INPUT FORM*/
-.userLocationDiv {
-  display: flex;
-  align-items: center;
-  border-radius: 15px;
-  background-color: var(--input-grey);
-}
-.userLocationDiv span {
-  margin: 0 auto;
-  padding: 5px 10px;
-}
-
-/* PREDICTIVE TEXT DROP DOWN */
-.locationPredictiveResults {
-  background-color: var(--input-grey);
-  z-index: 100;
-}
-
-.locationPredictiveResults ul {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.locationPredictiveResults li {
-  padding: 10px;
-  border-top: 1px rgb(194, 194, 194) solid;
-}
-
-.locationPredictiveResults li:hover {
-  background-color: var(--blue);
-  cursor: pointer;
-}
-
-.welcomePageDiv {
-  display: flex;
-  flex-direction: column;
-}
-.welcomePageDiv form {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-}
-.welcomePageDiv label {
-  font-family: var(--bodyFont);
-  font-weight: bold;
-  padding: var(--spacing);
-}
-
-.welcomePageDiv input {
-  width: 90%;
-  margin: 0 auto;
-  background-color: var(--input-grey);
-  border: none;
-  padding: var(--spacing);
-  border-radius: 15px;
-}
-
-.welcomePageDiv .wrapper {
-  display: flex;
-  flex-direction: column;
-}
-
-.subHeaderBox {
-  margin: 100px 0;
-}
-
-/* LOCATION FORM PAGE */
-
-.locationForm {
-  margin: 20px 0;
-}
-
-.locationForm p {
-  text-align: center;
-}
-
-.findLocation,
-.getStartedButton {
-  background: transparent;
-  border: none;
-}
-
-.findLocation,
-.getStartedButton a {
-  background-color: var(--blue);
-  font-family: var(--bodyFont);
-  transition: 0.3s background-color;
-  border: none;
-  padding: var(--spacing);
-  border-radius: 15px;
-  font-weight: bold;
-  color: white;
-}
-
-.findLocation:hover,
-.findLocation:focus,
-.getStartedButton a:hover,
-.getStartedButton a:focus {
-  cursor: pointer;
-  background-color: rgb(33, 52, 87);
-}
-
-.getStartedButton a {
-  text-decoration: none;
-  color: white;
-}
-
-.backButton {
-  background-color: transparent;
-  border: none;
-  margin: 20px;
-}
-
-.backButton:hover {
-  background-color: transparent;
-}
-
-.backButton a {
-  text-decoration: none;
-  font-family: var(--bodyFont);
-  color: black;
-  transition: 0.3s color;
-}
-
-.backButton a:hover,
-.backButton a:focus {
-  color: brown;
-}
-
-/* Loading animation */
-.loadingP {
-  font-family: var(--bodyFont);
-  font-size: clamp(1.2rem, 4vw, 1.5rem);
-}
-.sp {
-  width: 32px;
-  height: 32px;
-  clear: both;
-  margin: 20px auto;
-}
-
-.dot-flashing,
-.dot-flashing::before,
-.dot-flashing::after {
-  width: 50px;
-  height: 50px;
-  border-radius: 50px;
-  background-color: #5bc5ff;
-}
-
-.dot-flashing {
-  position: relative;
-  animation: dotFlashing 1s infinite linear alternate;
-  animation-delay: 0.5s;
-}
-
-.dot-flashing::before,
-.dot-flashing::after {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  top: 0;
-  animation: dotFlashing 1s infinite alternate;
-}
-
-.dot-flashing::before {
-  left: -60px;
-  animation-delay: 0s;
-}
-
-.dot-flashing::after {
-  left: 60px;
-  animation-delay: 1s;
-}
-
-@keyframes dotFlashing {
-  0% {
-    background-color: #c4e39c;
-  }
-  50% {
-    background-color: #5bc5ff;
-  }
-  100% {
-    background-color: #ff9d7f;
-  }
-}
