+++ src/Styles/Main.css
@@ -1,0 +1,67 @@
+/* Main section */
+
+main {
+  background-color: white;
+  border-top-left-radius: 15px;
+  border-top-right-radius: 15px;
+  height: 50vh;
+}
+
+/* Loading animation */
+.loadingP {
+  text-align: center;
+  font-family: var(--bodyFont);
+  font-size: clamp(1.2rem, 4vw, 1.5rem);
+}
+.sp {
+  width: 32px;
+  height: 32px;
+  clear: both;
+  margin: 20px auto;
+}
+
+.dot-flashing,
+.dot-flashing::before,
+.dot-flashing::after {
+  width: 50px;
+  height: 50px;
+  border-radius: 50px;
+  background-color: #5bc5ff;
+}
+
+.dot-flashing {
+  position: relative;
+  animation: dotFlashing 1s infinite linear alternate;
+  animation-delay: 0.5s;
+}
+
+.dot-flashing::before,
+.dot-flashing::after {
+  content: "";
+  display: inline-block;
+  position: absolute;
+  top: 0;
+  animation: dotFlashing 1s infinite alternate;
+}
+
+.dot-flashing::before {
+  left: -60px;
+  animation-delay: 0s;
+}
+
+.dot-flashing::after {
+  left: 60px;
+  animation-delay: 1s;
+}
+
+@keyframes dotFlashing {
+  0% {
+    background-color: #c4e39c;
+  }
+  50% {
+    background-color: #5bc5ff;
+  }
+  100% {
+    background-color: #ff9d7f;
+  }
+}
