+++ src/components/Header.js
@@ -1,7 +1,7 @@
-import homeImage0 from "./assets/store-1.png";
-import homeImage1 from "./assets/store-2.png";
-import homeImage2 from "./assets/store-3.png";
-import homeImage3 from "./assets/store-4.png";
+import homeImage0 from "../assets/store-1.png";
+import homeImage1 from "../assets/store-2.png";
+import homeImage2 from "../assets/store-3.png";
+import homeImage3 from "../assets/store-4.png";
 
 function Header() {
   const images = [homeImage0, homeImage1, homeImage2, homeImage3];
